+++ hoomd/CommunicatorGPU.cc
@@ -29,7 +29,8 @@ CommunicatorGPU::CommunicatorGPU(std::shared_ptr<SystemDefinition> sysdef,
       m_angle_comm(*this, m_sysdef->getAngleData()),
       m_dihedral_comm(*this, m_sysdef->getDihedralData()),
       m_improper_comm(*this, m_sysdef->getImproperData()),
-      m_constraint_comm(*this, m_sysdef->getConstraintData())
+      m_constraint_comm(*this, m_sysdef->getConstraintData()),
+      m_pair_comm(*this, m_sysdef->getPairData())
     {
     // default value
     #ifndef ENABLE_MPI_CUDA
@@ -1534,6 +1535,10 @@ void CommunicatorGPU::migrateParticles()
         m_bond_comm.migrateGroups(m_bonds_changed, true);
         m_bonds_changed = false;
 
+        // Special pairs
+        m_pair_comm.migrateGroups(m_pairs_changed, true);
+        m_pairs_changed = false;
+
         // Angles
         m_angle_comm.migrateGroups(m_angles_changed, true);
         m_angles_changed = false;
@@ -1866,6 +1871,9 @@ void CommunicatorGPU::exchangeGhosts()
 
         // bonds
         m_bond_comm.markGhostParticles(m_ghost_plan,m_comm_mask[stage]);
+
+        // special pairs
+        m_pair_comm.markGhostParticles(m_ghost_plan,m_comm_mask[stage]);
 
         // angles
         m_angle_comm.markGhostParticles(m_ghost_plan,m_comm_mask[stage]);