+++ hoomd/hpmc/test-py/stats_check.py
@@ -122,7 +122,7 @@ class pair_accept_some(unittest.TestCase):
     def setUp(self) :
         self.system  = create_empty(N=1000, box=data.boxdim(L=12, dimensions=3), particle_types=['A','B'])
 
-        self.mc = hpmc.integrate.ellipsoid(seed=10,d=1.0);
+        self.mc = hpmc.integrate.ellipsoid(seed=84,d=1.0);
         self.mc.shape_param.set('A', a=0.5,b=0.5,c=0.5,ignore_statistics=True)
         self.mc.overlap_checks.set('A','A', False)
         self.mc.shape_param.set('B', a=0.5,b=0.5,c=0.5,ignore_statistics=False)